requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.6
django-ckeditor==6.7.0
django-colorfield==0.10.1
django-import-export==3.3.1
django-jet-reboot==1.3.3
django-js-asset==2.1.0
django-resized==1.0.2
et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.4
mailchimp-marketing==3.0.80
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sorl-thumbnail==12.10.0
sqlparse==0.4.4
tablib==3.5.0
text-unidecode==1.3
tzdata==2023.3
urllib3==2.0.7
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
